ml/requirements.txt
pytorch
pyarrow
pandas
numpy
onnx